move post file to admin  , add index admin
