--- server/controllers/admin/getEventsDay.js
@@ -1,1 +1,0 @@
-module.exports = (req, res) => {};
