+++ server/controllers/admin/index.js
@@ -1,0 +1,4 @@
+const getEventsDay = require('./event/getEventsDay');
+const postEvent = require('./event/postEvent');
+
+module.exports = { getEventsDay, postEvent };
